It now uses haml as this seems easy to render and is a good chance to explore it.
Also uses chosen jQuery to narrow the page to only show the matching row (might alter this to show some around it).

Further improvements to design required, have added in ability to add additional stylesheets, with the aim of doing the same for images and javascript files.

Also changed it to use a new json that seems to be more concise and full.